+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Williams](https://togaware.com/Graham.Williams.html).
 
 ## 1.0 New Secure Key Handling
 
++ Add OVERDUE screen [1.0.16 20260726 gjw]
 + Update to latest solidui with cached profile [1.0.15 20260718 gjw]
 + BACKUP -> Export [1.0.14 20260717 gjw]
 + OIDC update for chrome/web [1.0.13 20260713 tonypioneer]